loading_DA64.c FUN_8013234c

+++ include/loading_DA64.h
@@ -26,7 +26,7 @@ extern AllFixData *PS1_AllFixData;
 
 void PS1_LoadLevelMapBlock(MapData *data);
 void FUN_80132304(u32 *data, u8 height);
-void FUN_8013234c(s32 *param_1);
+void FUN_8013234c(u8 *param_1);
 void FUN_80132424(void);
 void PS1_LoadAllFixData(void);
 void PS1_LoadLevelEventBlock(void);
